+++ ChangeLog.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## 2.2
+
+- FIX: PHP 8 - *04/08/2022* - 2.2.3
+- FIX: Compatibility V16 - newToken and family - *06/06/2022* - 2.2.2
+- FIX: Stock's column colour correction on order - *2021-10-06* - 2.2.1
+- NEW: Colonne stock virtuel - *2021-09-21* - 2.2.0
+
 ## 2.1
 
 - FIX: v13 — 14 compatibility: societe_id,  - *2021-09-09* - 2.1.3